retrieval-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
faiss-cpu==1.9.0
sentence-transformers==3.3.1
numpy==2.1.3
pydantic==2.10.3